requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
